refactoring. 모듈 구조 분리 및 pay 모듈 추가

+++ api/src/test/java/com/commerce/saleday/api/service/orchestrator/ItemStockOrchestratorServiceTest.java
@@ -1,10 +1,11 @@
 package com.commerce.saleday.api.service.orchestrator;
 
-import com.commerce.saleday.api.service.item.ItemService;
-import com.commerce.saleday.domain.item.model.Item;
-import com.commerce.saleday.domain.review.model.Review;
-import com.commerce.saleday.domain.stock.model.ItemStock;
+import com.commerce.saleday.order.service.ItemService;
+import com.commerce.saleday.order.domain.item.model.Item;
+import com.commerce.saleday.order.domain.review.model.Review;
+import com.commerce.saleday.order.domain.stock.model.ItemStock;
 import jakarta.transaction.Transactional;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
@@ -30,7 +31,7 @@ class ItemStockOrchestratorServiceTest {
     String code = "1235";
     String name = "테스트과자";
     String content = "테스트달달한과자";
-    double price = 10000;
+    BigDecimal price = BigDecimal.valueOf(10000);
     List<Review> reviews = new ArrayList<>();
 
     Item item = Item.create(code, name, content, price, reviews);
